# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WALAB API Documentation",
        "description": "Documentation technique complète pour l'application Walab. Utilisez le bouton 'Authorize' pour tester les routes protégées.",
        "contact": {
            "email": "contact@walab.bj"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur API Principal"
        }
    ],
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Inscription d'un nouvel utilisateur",
                "description": "Permet d'inscrire un Patient ou un Professionnel de santé.",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstname",
                                    "lastname",
                                    "email",
                                    "password",
                                    "phone",
                                    "user_type"
                                ],
                                "properties": {
                                    "firstname": {
                                        "type": "string",
                                        "example": "Jean"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "example": "Dupont"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jean.dupont@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password123!"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+22901020304"
                                    },
                                    "adress": {
                                        "type": "string",
                                        "example": "Cotonou, Quartier Jak"
                                    },
                                    "ville": {
                                        "type": "string",
                                        "example": "Cotonou"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "Masculin",
                                            "Féminin"
                                        ],
                                        "example": "Masculin"
                                    },
                                    "date_naissance": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1995-05-20"
                                    },
                                    "user_type": {
                                        "description": "Type de compte",
                                        "type": "string",
                                        "enum": [
                                            "client",
                                            "professionnel"
                                        ],
                                        "example": "client"
                                    },
                                    "order_number": {
                                        "description": "Requis si user_type = professionnel",
                                        "type": "string",
                                        "example": "MED-8392"
                                    },
                                    "profession": {
                                        "description": "Requis si user_type = professionnel",
                                        "type": "string",
                                        "example": "Médecin Généraliste"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inscription réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur inscrit avec succès"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|AbCdEf123..."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation (Email existant, champs manquants...)"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion utilisateur",
                "description": "Authentification par email et mot de passe.",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "client@walab.bj"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Résultat de la tentative de connexion",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "description": "Connexion réussie",
                                            "properties": {
                                                "success": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Utilisateur connecté"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "2|XyZaBc..."
                                                },
                                                "data": {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "description": "Email non vérifié ou compte inactif",
                                            "properties": {
                                                "success": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Veuillez vérifier votre adresse email."
                                                },
                                                "requires_verification": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "user@mail.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants incorrects (si le code retourne 401, sinon 200 avec success=false)"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion",
                "description": "Révoque le token d'authentification actuel.",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur déconnecté"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "Utilisateur",
                "description": "Modèle utilisateur standard",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "firstname": {
                        "type": "string",
                        "example": "Jean"
                    },
                    "lastname": {
                        "type": "string",
                        "example": "Dupont"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jean@mail.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+22901020304"
                    },
                    "role_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrez votre token d'accès ici (sans le préfixe 'Bearer')",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}